Removing old job manager API
