--- components/org.wso2.carbon.das.jobmanager.core/src/gen/java/io/swagger/api/Application.java
@@ -1,16 +1,0 @@
-package io.swagger.api;
-
-import org.wso2.msf4j.MicroservicesRunner;
-
-/**
- * Application entry point.
- *
- * @since 1.0.0-SNAPSHOT
- */
-public class Application {
-    public static void main(String[] args) {
-        new MicroservicesRunner()
-                .deploy(new ResourceManagerApi())
-                .start();
-    }
-}
