--- components/org.wso2.carbon.das.jobmanager.core/src/gen/java/io/swagger/api/JacksonJsonProvider.java
@@ -1,27 +1,0 @@
-package io.swagger.api;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.joda.JodaModule;
-import com.fasterxml.jackson.jaxrs.json.JacksonJaxbJsonProvider;
-
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.ext.Provider;
-
-@Provider
-@Produces({MediaType.APPLICATION_JSON})
-public class JacksonJsonProvider extends JacksonJaxbJsonProvider {
-
-    public JacksonJsonProvider() {
-
-        ObjectMapper objectMapper = new ObjectMapper()
-                .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
-                .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-                .registerModule(new JodaModule())
-                .setDateFormat(new RFC3339DateFormat());
-
-        setMapper(objectMapper);
-    }
-}
