+++ components/org.wso2.carbon.das.jobmanager.core/src/gen/java/org/wso2/carbon/das/jobmanager/core/ApiException.java
@@ -1,26 +1,13 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
 package org.wso2.carbon.das.jobmanager.core;
 
-@javax.annotation.Generated(value = "org.wso2.maven.plugins.JavaMSF4JServerCodegen", date = "2017-03-14T10:36:41.439+05:30")
-public class ApiException extends Exception{
+@javax.annotation.Generated(
+        value = "io.swagger.codegen.languages.JavaMSF4JServerCodegen",
+        date = "2017-10-17T16:36:16.751Z"
+)
+public class ApiException extends Exception {
     private int code;
-    public ApiException (int code, String msg) {
+
+    public ApiException(int code, String msg) {
         super(msg);
         this.code = code;
     }
