+++ components/org.wso2.carbon.das.jobmanager.core/src/gen/java/org/wso2/carbon/das/jobmanager/core/StringUtil.java
@@ -1,59 +1,45 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
 package org.wso2.carbon.das.jobmanager.core;
 
-@javax.annotation.Generated(value = "org.wso2.maven.plugins.JavaMSF4JServerCodegen", date = "2017-03-14T10:36:41.439+05:30")
+@javax.annotation.Generated(
+        value = "io.swagger.codegen.languages.JavaMSF4JServerCodegen",
+        date = "2017-10-17T16:36:16.751Z"
+)
 public class StringUtil {
-  /**
-   * Check if the given array contains the given value (with case-insensitive comparison).
-   *
-   * @param array The array
-   * @param value The value to search
-   * @return true if the array contains the value
-   */
-  public static boolean containsIgnoreCase(String[] array, String value) {
-    for (String str : array) {
-      if (value == null && str == null) return true;
-      if (value != null && value.equalsIgnoreCase(str)) return true;
+    /**
+     * Check if the given array contains the given value (with case-insensitive comparison).
+     *
+     * @param array The array
+     * @param value The value to search
+     * @return true if the array contains the value
+     */
+    public static boolean containsIgnoreCase(String[] array, String value) {
+        for (String str : array) {
+            if (value == null && str == null) return true;
+            if (value != null && value.equalsIgnoreCase(str)) return true;
+        }
+        return false;
     }
-    return false;
-  }
 
-  /**
-   * Join an array of strings with the given separator.
-   * <p>
-   * Note: This might be replaced by utility method from commons-lang or guava someday
-   * if one of those libraries is added as dependency.
-   * </p>
-   *
-   * @param array     The array of strings
-   * @param separator The separator
-   * @return the resulting string
-   */
-  public static String join(String[] array, String separator) {
-    int len = array.length;
-    if (len == 0) return "";
+    /**
+     * Join an array of strings with the given separator.
+     * <p>
+     * Note: This might be replaced by utility method from commons-lang or guava someday
+     * if one of those libraries is added as dependency.
+     * </p>
+     *
+     * @param array     The array of strings
+     * @param separator The separator
+     * @return the resulting string
+     */
+    public static String join(String[] array, String separator) {
+        int len = array.length;
+        if (len == 0) return "";
 
-    StringBuilder out = new StringBuilder();
-    out.append(array[0]);
-    for (int i = 1; i < len; i++) {
-      out.append(separator).append(array[i]);
+        StringBuilder out = new StringBuilder();
+        out.append(array[0]);
+        for (int i = 1; i < len; i++) {
+            out.append(separator).append(array[i]);
+        }
+        return out.toString();
     }
-    return out.toString();
-  }
 }
