--- components/org.wso2.carbon.das.jobmanager.core/src/gen/java/org/wso2/carbon/das/jobmanager/core/dto/ErrorDTO.java
@@ -1,188 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.wso2.carbon.das.jobmanager.core.dto;
-
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModelProperty;
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * ErrorDTO
- */
-@javax.annotation.Generated(value = "org.wso2.maven.plugins.JavaMSF4JServerCodegen", date = "2017-03-14T10:36:41.439+05:30")
-public class ErrorDTO   {
-  @JsonProperty("code")
-  private Long code = null;
-
-  @JsonProperty("message")
-  private String message = null;
-
-  @JsonProperty("description")
-  private String description = null;
-
-  @JsonProperty("moreInfo")
-  private String moreInfo = null;
-
-  @JsonProperty("error")
-  private List<ErrorListItemDTO> error = new ArrayList<ErrorListItemDTO>();
-
-  public ErrorDTO code(Long code) {
-    this.code = code;
-    return this;
-  }
-
-   /**
-   * Get code
-   * @return code
-  **/
-  @ApiModelProperty(required = true, value = "")
-  public Long getCode() {
-    return code;
-  }
-
-  public void setCode(Long code) {
-    this.code = code;
-  }
-
-  public ErrorDTO message(String message) {
-    this.message = message;
-    return this;
-  }
-
-   /**
-   * Error message.
-   * @return message
-  **/
-  @ApiModelProperty(required = true, value = "Error message.")
-  public String getMessage() {
-    return message;
-  }
-
-  public void setMessage(String message) {
-    this.message = message;
-  }
-
-  public ErrorDTO description(String description) {
-    this.description = description;
-    return this;
-  }
-
-   /**
-   * A detail description about the error message. 
-   * @return description
-  **/
-  @ApiModelProperty(value = "A detail description about the error message. ")
-  public String getDescription() {
-    return description;
-  }
-
-  public void setDescription(String description) {
-    this.description = description;
-  }
-
-  public ErrorDTO moreInfo(String moreInfo) {
-    this.moreInfo = moreInfo;
-    return this;
-  }
-
-   /**
-   * Preferably an url with more details about the error. 
-   * @return moreInfo
-  **/
-  @ApiModelProperty(value = "Preferably an url with more details about the error. ")
-  public String getMoreInfo() {
-    return moreInfo;
-  }
-
-  public void setMoreInfo(String moreInfo) {
-    this.moreInfo = moreInfo;
-  }
-
-  public ErrorDTO error(List<ErrorListItemDTO> error) {
-    this.error = error;
-    return this;
-  }
-
-  public ErrorDTO addErrorItem(ErrorListItemDTO errorItem) {
-    this.error.add(errorItem);
-    return this;
-  }
-
-   /**
-   * If there are more than one error list them out. For example, list out validation errors by each field. 
-   * @return error
-  **/
-  @ApiModelProperty(value = "If there are more than one error list them out. For example, list out validation errors by each field. ")
-  public List<ErrorListItemDTO> getError() {
-    return error;
-  }
-
-  public void setError(List<ErrorListItemDTO> error) {
-    this.error = error;
-  }
-
-
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    ErrorDTO error = (ErrorDTO) o;
-    return Objects.equals(this.code, error.code) &&
-        Objects.equals(this.message, error.message) &&
-        Objects.equals(this.description, error.description) &&
-        Objects.equals(this.moreInfo, error.moreInfo) &&
-        Objects.equals(this.error, error.error);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(code, message, description, moreInfo, error);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class ErrorDTO {\n");
-    
-    sb.append("    code: ").append(toIndentedString(code)).append("\n");
-    sb.append("    message: ").append(toIndentedString(message)).append("\n");
-    sb.append("    description: ").append(toIndentedString(description)).append("\n");
-    sb.append("    moreInfo: ").append(toIndentedString(moreInfo)).append("\n");
-    sb.append("    error: ").append(toIndentedString(error)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(java.lang.Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
